requirements.txt
python-dotenv
groq
streamlit
sentence-transformers
requests
watchdog
numpy